+++ src/components/Tabs.tsx
@@ -1,0 +1,58 @@
+import classNames from 'classnames';
+import { Link, useParams } from 'react-router-dom';
+import { Tab } from '../types/Tab';
+
+const tabs: Tab[] = [
+  {
+    id: 'tab-1',
+    title: 'Tab 1',
+    content: 'Some text 1',
+  },
+  {
+    id: 'tab-2',
+    title: 'Tab 2',
+    content: 'Some text 2',
+  },
+  {
+    id: 'tab-3',
+    title: 'Tab 3',
+    content: 'Some text 3',
+  },
+];
+
+export const Tabs = () => {
+  const { tabId } = useParams();
+
+  const selectedTab = tabs.find(tab => tab.id === tabId);
+
+  return (
+    <>
+      <div className="tabs is-boxed">
+        <ul>
+          {tabs.map(tab => (
+            <li
+              key={tab.id}
+              data-cy="Tab"
+              className={classNames({
+                'is-active': tab.id === tabId,
+              })}
+            >
+              <Link to={`/tabs/${tab.id}`}>
+                {tab.title}
+              </Link>
+            </li>
+          ))}
+        </ul>
+      </div>
+
+      <div
+        className="block"
+        data-cy="TabContent"
+      >
+        {selectedTab
+          ? selectedTab.content
+          : 'Please select a tab'}
+      </div>
+    </>
+  );
+};
